+++ tools/cer_log_joint/main.cpp
@@ -62,7 +62,7 @@ public:
         string robot=rf.check("robot",Value("cer")).asString();
         string part=rf.check("part",Value("left_upper_arm")).asString();
         type=rf.check("type",Value("stairs")).asString();
-        joint=rf.check("joint",Value(0)).asInt();
+        joint=rf.check("joint",Value(0)).asInt8();
         Ts=rf.check("Ts",Value(0.01)).asFloat64();
         T=rf.check("T",Value(4.0)).asFloat64();
